refactor: extract fetchElementCountByStatus to shared insights helper
Removes ProcessInstancesService dependency from CaseInstancesService.
Both services now call the shared function directly via their own
BaseService.post method.
